Player fixes and other bug-fixes
I made it so the object component was a but more useful so now, when an object has an object component it automatically sets it up to be interacted with and also has better 'gravity'. Do not worry, the buttons did not need the object components to work and hadn't for a while so that's all good there. I also got the big player moving objects work much better and the web shoot too. There were some other small adjustments too like giving some more objects the objectComponent script so there is a bit more interactivity